backend/requirements.txt
flask
python-dotenv
psycopg2-binary
passlib
flask_cors
pandas
newsapi
bs4
urllib
time
requests